# frontend/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Radar API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/ai/signals/{signal_id}/explanations": {
      "get": {
        "tags": [
          "ai"
        ],
        "summary": "List Signal Ai Explanations",
        "operationId": "list_signal_ai_explanations_api_v1_ai_signals__signal_id__explanations_get",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalAIExplanationResponse"
                  },
                  "title": "Response List Signal Ai Explanations Api V1 Ai Signals  Signal Id  Explanations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/signals/{signal_id}/explanations/generate": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Generate Signal Ai Explanation",
        "operationId": "generate_signal_ai_explanation_api_v1_ai_signals__signal_id__explanations_generate_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SignalAIExplanationGenerateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalAIExplanationResponse"
                }
              }
            }
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIExplanationNotReadyResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtests/run": {
      "post": {
        "tags": [
          "backtests"
        ],
        "summary": "Run Backtest",
        "operationId": "run_backtest_api_v1_backtests_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestRunResult"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestNotReadyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtests/results": {
      "get": {
        "tags": [
          "backtests"
        ],
        "summary": "List Backtest Results",
        "operationId": "list_backtest_results_api_v1_backtests_results_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BacktestResultResponse"
                  },
                  "title": "Response List Backtest Results Api V1 Backtests Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/subscription": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Billing Subscription",
        "operationId": "get_billing_subscription_api_v1_billing_subscription_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo_user",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Billing Checkout",
        "operationId": "create_billing_checkout_api_v1_billing_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderActionResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderNotReadyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/customer-portal": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Billing Customer Portal",
        "operationId": "create_billing_customer_portal_api_v1_billing_customer_portal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillingPortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderActionResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderNotReadyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/webhooks/{provider}": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Handle Billing Webhook",
        "operationId": "handle_billing_webhook_api_v1_billing_webhooks__provider__post",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BillingWebhookRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderActionResponse"
                }
              }
            }
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingProviderNotReadyResponse"
                }
              }
            },
            "description": "Not Implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIExplanationNotReadyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "not_implemented",
            "title": "Status",
            "default": "not_implemented"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "signal_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Id"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "storage_target": {
            "type": "string",
            "title": "Storage Target",
            "default": "signal_ai_explanations"
          },
          "orchestrator_required": {
            "type": "boolean",
            "title": "Orchestrator Required",
            "default": true
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message",
          "model_provider",
          "model_name"
        ],
        "title": "AIExplanationNotReadyResponse"
      },
...
      "BacktestNotReadyResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "not_implemented",
            "title": "Status",
            "default": "not_implemented"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "worker_required": {
            "type": "boolean",
            "title": "Worker Required",
            "default": true
          },
          "analytics_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Analytics Targets"
          },
          "data_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "BacktestNotReadyResponse"
      },
      "BacktestResultResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "strategy_code": {
            "type": "string",
            "title": "Strategy Code"
          },
          "strategy_version": {
            "type": "string",
            "title": "Strategy Version"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "initial_capital": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Initial Capital"
          },
          "final_equity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Final Equity"
          },
          "pnl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pnl"
          },
          "pnl_pct": {
            "type": "number",
            "title": "Pnl Pct"
          },
          "max_drawdown_pct": {
            "type": "number",
            "title": "Max Drawdown Pct"
          },
          "trades_count": {
            "type": "integer",
            "title": "Trades Count"
          },
          "wins_count": {
            "type": "integer",
            "title": "Wins Count"
          },
          "losses_count": {
            "type": "integer",
            "title": "Losses Count"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "equity_curve": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Equity Curve"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "user_id",
          "strategy_code",
          "strategy_version",
          "exchange",
          "symbol",
          "timeframe",
          "start_at",
          "end_at",
          "initial_capital",
          "final_equity",
          "pnl",
          "pnl_pct",
          "max_drawdown_pct",
          "trades_count",
          "wins_count",
          "losses_count",
          "created_at"
        ],
        "title": "BacktestResultResponse"
      },
      "BacktestRunRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "strategy_code": {
            "type": "string",
            "minLength": 1,
            "title": "Strategy Code"
          },
          "strategy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Version"
          },
          "exchange": {
            "type": "string",
            "minLength": 1,
            "title": "Exchange"
          },
          "symbol": {
            "type": "string",
            "minLength": 1,
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "minLength": 1,
            "title": "Timeframe"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "initial_capital": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Initial Capital",
            "default": "1000"
          },
          "fee_rate": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Fee Rate",
            "default": "0.001"
          },
          "slippage_bps": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Slippage Bps",
            "default": "0"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "strategy_code",
          "exchange",
          "symbol",
          "timeframe",
          "start_at",
          "end_at"
        ],
        "title": "BacktestRunRequest"
      },
      "BacktestRunResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "queued"
            ],
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestResultResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "BacktestRunResult"
      },
      "BillingCheckoutRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "plan_code": {
            "type": "string",
            "minLength": 1,
            "title": "Plan Code"
          },
          "success_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Url"
          },
          "cancel_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "plan_code"
        ],
        "title": "BillingCheckoutRequest"
      },
...
      "BillingProviderActionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "created",
              "handled"
            ],
            "title": "Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "status",
          "provider"
        ],
        "title": "BillingProviderActionResponse"
      },
...
      "BillingSubscriptionResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "trialing",
              "past_due",
              "canceled",
              "none"
            ],
            "title": "State"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "plan_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Code"
          },
          "plan_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Name"
          },
          "current_period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "external_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Provider"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits"
          },
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "state",
          "tier",
          "plan_id",
          "plan_code",
          "plan_name",
          "current_period_start",
          "current_period_end",
          "external_provider",
          "external_id",
          "limits",
          "features"
        ],
        "title": "BillingSubscriptionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignalAIExplanationGenerateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "demo_user"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider",
            "default": "stub"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": "not-configured"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "title": "SignalAIExplanationGenerateRequest"
      },
      "SignalAIExplanationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "signal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signal Id"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "prompt_hash": {
            "type": "string",
            "title": "Prompt Hash"
          },
          "explanation_md": {
            "type": "string",
            "title": "Explanation Md"
          },
          "risk_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "signal_id",
          "model_provider",
          "model_name",
          "prompt_hash",
          "explanation_md",
          "risk_notes",
          "created_at"
        ],
        "title": "SignalAIExplanationResponse"
      },
...
    }
  }
}